+++ .gitignore
@@ -1,0 +1,5 @@
+__pycache__/**
+.DS_Store
+.ipynb_checkpoints
+.ipynb_checkpoints/**
+original_modules/__pycache__/**
